requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.12
djangorestframework==3.16.1
exceptiongroup==1.3.1
idna==3.11
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
redis==7.3.0
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
